+++ lib/card_dealer.rb
@@ -2,10 +2,12 @@
 
 require "set"
 
-require_relative "card_dealer/version"
-require_relative "card_dealer/card"
-require_relative "card_dealer/deck"
-
 module CardDealer
   class Error < StandardError; end
 end
+
+require_relative "card_dealer/version"
+require_relative "card_dealer/card"
+require_relative "card_dealer/deck"
+require_relative "card_dealer/deck_builder"
+require_relative "card_dealer/binary_deck"